requirements.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
idna==2.10
pycodestyle==2.7.0
python-dateutil==2.8.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.5
